Fully functioned auth and added docs

+++ app/Http/Controllers/Api/V1/Admin/RespondenApiController.php
@@ -59,12 +59,15 @@ class RespondenApiController extends Controller
      *         )
      *      ),
      *     @OA\Response(response="403", description="Unauthorized"),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      *   ),
      */
 
     public function index()
     {
-        // abort_if(Gate::denies('responden_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('responden_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return new RespondenResource(Responden::all());
     }
@@ -124,14 +127,17 @@ class RespondenApiController extends Controller
      *          @OA\JsonContent(
      *          type="object",
      *          @OA\Property(property="data", type="object",
-     *              @OA\Property(property="nama", type="string", example="Jimans"),
-     *              @OA\Property(property="kode", type="string", example="332"),
-     *              @OA\Property(property="usia", type="integer", example=19),
-     *              @OA\Property(property="paritas", type="integer", example=null),
-     *              @OA\Property(property="kardiotografi", type="integer", example=null),
-     *              @OA\Property(property="created_at", type="string", example="2023-08-28 20:55:36"),
-     *              @OA\Property(property="updated_at", type="string", example="2023-08-28 20:55:36"),
-     *              @OA\Property(property="id", type="integer", example=1),
+     *                      @OA\Property(property="id", type="integer", example=1),
+     *                      @OA\Property(property="nama", type="string", example="Jimans"),
+     *                      @OA\Property(property="kode", type="string", example="334"),
+     *                      @OA\Property(property="usia", type="integer", example=20),
+     *                      @OA\Property(property="his_adekuat", type="integer", example=null),
+     *                      @OA\Property(property="pergerakan", type="integer", example=null),
+     *                      @OA\Property(property="paritas", type="integer", example=null),
+     *                      @OA\Property(property="alamat", type="integer", example=null),
+     *                      @OA\Property(property="created_at", type="string", example="2023-08-28 20:55:36"),
+     *                      @OA\Property(property="updated_at", type="string", example="2023-08-28 14:51:56"),
+     *                      @OA\Property(property="deleted_at", type="integer", example=null),
      *         )
      *     )
      *      ),
@@ -143,17 +149,20 @@ class RespondenApiController extends Controller
      *      ),
      *     @OA\Response(response="403", description="Unauthorized"),
      *     @OA\Response(response="422", description="Missing field",
-     *     @OA\JsonContent(
-     *         type="object",
-     *         @OA\Property(property="message", type="string", example="The usia field is required."),
-     *         @OA\Property(property="errors", type="object",
-     *            @OA\Property(property="usia", type="array",
-     *              @OA\Items(type="string", example="The usia field is required.")
-     *            ),
-     *         )
-     *     ),
+     *          @OA\JsonContent(
+     *              type="object",
+     *              @OA\Property(property="message", type="string", example="The usia field is required."),
+     *              @OA\Property(property="errors", type="object",
+     *                  @OA\Property(property="usia", type="array",
+     *                      @OA\Items(type="string", example="The usia field is required.")
+     *                  ),
+     *              )
+     *          ),
      *    
      *     ),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      *   )
      */
 
@@ -210,12 +219,15 @@ class RespondenApiController extends Controller
      *     @OA\Response(response="401", description="Unauthenticated"),
      *     @OA\Response(response="403", description="Unauthorized"),
      *     @OA\Response(response="404", description="Data not found"),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      *   ),
      */
 
     public function show(Responden $responden)
     {
-        // abort_if(Gate::denies('responden_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('responden_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return new RespondenResource($responden);
     }
@@ -321,7 +333,9 @@ class RespondenApiController extends Controller
      *         )
      *      ),
      *     @OA\Response(response=500, description="Internal Server Error"),
-     *     
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      *    ),
      *  @OA\Patch(
      *     tags={"Respondens"},
@@ -424,7 +438,9 @@ class RespondenApiController extends Controller
      *         )
      *      ),
      *     @OA\Response(response=500, description="Internal Server Error"),
-     *     
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      *    ),
      *  ),
      */
@@ -466,14 +482,16 @@ class RespondenApiController extends Controller
      *     ),
      *     @OA\Response(response="401", description="Unauthenticated"),
      *     @OA\Response(response="403", description="Unauthorized"),
-     *     @OA\Response(response="404", description="Data not found"
-     * ),
+     *     @OA\Response(response="404", description="Data not found"),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
      * ),
      */
 
     public function destroy(Responden $responden)
     {
-        // abort_if(Gate::denies('responden_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('responden_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $responden->delete();
